backend/requirements.txt
# ===========================================
# Web Compliance Checker - Backend Dependencies
# ===========================================

# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate limiting
slowapi==0.1.9

# Browser automation (Playwright)
playwright==1.40.0

# AI/ML and API clients
replicate==0.25.0

# DNS resolution for security checks
dnspython==2.4.2

# Environment variables
python-dotenv==1.0.0

# HTTP client for health checks
httpx==0.25.2
